Drag and drop to the drop area didn't open new tab for me, but dropping it outside the drop area did. Fixed that in imageInput.js with document listeners for drop and dragover events to prevent defaults.
Clicking cancel in File Explorer triggered the "Please upload a valid image file" so I added a check for empty file in onImageSelected to fix that.

Upload image header was displaying over images passed into the ImageInput component, so I gave it a dynamic id in the custom component and set it so the display would set to none when a valid image was uploaded.