Store access when adding and updating users. Navigating away from the page (or refreshing) can disrupt data transmission (sometimes resulting in name, email, and role being added without store access), so a Loading Message modal is implemented to warn the user from doing that when adding a user.
When selecting and moving stores, it reliable moves stores when Ctrl-Click is used to highlight the stores (You have to have them highlighted, if at first the store is not highlighted but in a dashed box you need to click it again). It works inconsistently otherwise. Paragraph element informs user to use Ctrl-Click.